#144 Garbage collection des fichiers du datastore [fixed #144] Ajout du AttachementJobService et de l'action sur la page de maintenance

+++ eliot-tdbase-plugin/grails-app/services/org/lilie/services/eliot/tdbase/CopieJobService.groovy
@@ -86,7 +86,7 @@ class CopieJobService {
 }
 
 class SupprimeCopiesJetablesRapport {
-  Integer nombreDeCopiesSupprimees = 0
-  Integer nombreDeReponsesSupprimees = 0
+  Long nombreDeCopiesSupprimees = 0
+  Long nombreDeReponsesSupprimees = 0
   Date dateDernierEnregistrementCopiesSupprimees
 }
